RecognizeNumber.asv

+++ Statistical Charts/getRectArea.m
@@ -1,6 +1,6 @@
 function [ area ] =  getRectArea(maxx, minx,maxy,miny)
 %UNTITLED4 Summary of this function goes here
 %   Detailed explanation goes here
-area = (uint16( maxx -minx ) *( uint16(maxy - miny ) ))  ;  
+area = (uint32( maxx -minx ) *( uint32(maxy - miny ) ))  ;  
 end
